perf: 新增自定义ACME管理功能，完善证书颁发机构适配

+++ packages/ui/certd-server/db/migration-pg/v10052__plugin_market_columns.sql
@@ -1,11 +1,11 @@
-ALTER TABLE "pi_plugin" ADD COLUMN "app_id" integer;
-ALTER TABLE "pi_plugin" ADD COLUMN "developer_id" integer;
+ALTER TABLE "pi_plugin" ADD COLUMN "app_id" bigint;
+ALTER TABLE "pi_plugin" ADD COLUMN "developer_id" bigint;
 ALTER TABLE "pi_plugin" ADD COLUMN "full_name" varchar(191);
 ALTER TABLE "pi_plugin" ADD COLUMN "latest" varchar(100);
 ALTER TABLE "pi_plugin" ADD COLUMN "status" varchar(100);
-ALTER TABLE "pi_plugin" ADD COLUMN "download_count" integer;
+ALTER TABLE "pi_plugin" ADD COLUMN "download_count" bigint;
 ALTER TABLE "pi_plugin" ADD COLUMN "score" real;
-ALTER TABLE "pi_plugin" ADD COLUMN "sync_time" integer;
+ALTER TABLE "pi_plugin" ADD COLUMN "sync_time" bigint;
 
 CREATE UNIQUE INDEX "index_plugin_full_name" ON "pi_plugin" ("full_name");
 CREATE INDEX "index_plugin_plugin_type" ON "pi_plugin" ("pluginType");
